refactor(siu): move in Pipeable, singledispatch2 from verbs
